fix: move collector schema into server directory
