+++ src/lib/server/db/schema/messages.ts
@@ -7,7 +7,7 @@ import { patients } from './patients';
 import { responses } from './responses';
 import { users } from './users';
 
-import type { CollectorSchemaType } from '$lib/schemas/collector';
+import type { CollectorSchemaType } from '$server/utils/collector';
 
 export const messages = mysqlTable('messages', {
   ...identifiable,
